U.5: cut read and list over to sqlite metadata
