+++ crates/atm-core/src/read/mod.rs
@@ -1,4 +1,6 @@
 pub(crate) mod filters;
+pub(crate) mod legacy_path;
+pub(crate) mod metadata_selection;
 pub(crate) mod seen_state;
 pub(crate) mod state;
 pub(crate) mod wait;
@@ -7,26 +9,27 @@ use std::collections::HashMap;
 use std::path::{Path, PathBuf};
 
 use serde::{Deserialize, Serialize};
-use serde_json::Value;
-use tracing::debug;
 
 use crate::address::AgentAddress;
 use crate::boundary;
 use crate::config;
 use crate::error::AtmError;
 use crate::identity;
-use crate::mailbox::source::{SourceFile, SourcedMessage, resolve_target};
-use crate::mailbox::surface::dedupe_message_id_surface;
+use crate::mailbox::source::{SourceFile, resolve_target};
 use crate::observability::{CommandEvent, ObservabilityPort};
 use crate::schema::{AtmMessageId, MessageEnvelope};
 use crate::service_runtime::{LocalServiceRuntime, RetainedServiceRuntime};
 use crate::service_runtime_store::{RetainedMailboxRuntime, legacy_runtime};
-use crate::threading::{ThreadIndex, is_ephemeral, is_expired_ephemeral};
+use crate::threading::ThreadIndex;
 use crate::types::{
     AckActivationMode, AgentName, CommandAction, DisplayBucket, IsoTimestamp, MessageClass,
     ReadSelection, SourceIndex, TaskId, TeamName,
 };
 use crate::workflow;
+use legacy_path::read_mail_legacy_path;
+use metadata_selection::{
+    effective_display_envelope, selection_state_for_mailbox_metadata_rows, sort_and_limit_selected,
+};
 
 pub const MAX_CONTAINS_FILTER_LEN: usize = 1024;
 pub const MAX_TIMEOUT_SECS: u64 = 3600;
@@ -192,7 +195,16 @@ pub fn read_mail(
     observability: &dyn ObservabilityPort,
 ) -> Result<ReadOutcome, AtmError> {
     let runtime = legacy_runtime();
-    read_mail_with_runtime(query, observability, &runtime)
+    let context = resolve_read_context(&query, &runtime)?;
+    read_mail_legacy_path(
+        query,
+        observability,
+        &runtime,
+        context.actor,
+        context.actor_team,
+        context.target,
+        context.seen_watermark,
+    )
 }
 
 pub fn read_mail_with_runtime(
@@ -208,60 +220,19 @@ fn read_mail_with_runtime_impl<R: RetainedServiceRuntime + RetainedMailboxRuntim
     observability: &dyn ObservabilityPort,
     runtime: &R,
 ) -> Result<ReadOutcome, AtmError> {
-    let config = runtime.load_config(&query.current_dir)?;
-    let actor = identity::resolve_actor_identity(query.actor_override.as_deref(), config.as_ref())?;
-    let actor_team = config::resolve_team(query.team_override.as_deref(), config.as_ref());
-    let target = resolve_target(
-        query.target_address.as_ref(),
-        &actor,
-        query.team_override.as_ref(),
-        config.as_ref(),
-    )?;
-
-    let team_dir = runtime.team_dir(&query.home_dir, &target.team)?;
-    if !team_dir.exists() {
-        return Err(AtmError::team_not_found(&target.team).with_recovery(
-            "Create the team config for the requested team or target a different team before retrying `atm read`.",
-        ));
-    }
-
-    let team_config = runtime.load_team_config(&team_dir)?;
-    if target.explicit
-        && !team_config
-            .members
-            .iter()
-            .any(|member| member.name == target.agent.as_str())
-    {
-        return Err(
-            AtmError::agent_not_found(&target.agent, &target.team).with_recovery(
-                "Update the team membership in config.json or read a different mailbox target.",
-            ),
-        );
-    }
-
+    let ReadRuntimeContext {
+        actor,
+        actor_team,
+        target,
+        seen_watermark,
+    } = resolve_read_context(&query, runtime)?;
     let own_inbox = actor == target.agent && actor_team.as_deref() == Some(target.team.as_str());
-    let seen_watermark = if query.seen_state_filter && query.selection_mode != ReadSelection::All {
-        runtime.load_seen_watermark(&query.home_dir, &target.team, &target.agent)?
-    } else {
-        None
-    };
     let mut metadata_rows =
         runtime.query_mailbox_metadata_rows(&query.home_dir, &target.team, &target.agent, None)?;
-    let has_legacy_keys = metadata_rows
+    if metadata_rows
         .iter()
-        .any(|row| row.message_key.as_ref().starts_with("legacy:"));
-    if has_legacy_keys && runtime.allows_legacy_mailbox_files() {
-        return read_mail_legacy_path(
-            query,
-            observability,
-            runtime,
-            actor,
-            actor_team,
-            target,
-            seen_watermark,
-        );
-    }
-    if has_legacy_keys {
+        .any(|row| row.message_key.as_ref().starts_with("legacy:"))
+    {
         return Err(AtmError::validation(
             "sqlite mailbox metadata returned legacy-prefixed message keys in non-legacy runtime mode",
         )
@@ -428,582 +399,60 @@ fn read_mail_with_runtime_impl<R: RetainedServiceRuntime + RetainedMailboxRuntim
     Ok(outcome)
 }
 
-fn read_mail_legacy_path<R: RetainedServiceRuntime + RetainedMailboxRuntime>(
-    query: ReadQuery,
-    observability: &dyn ObservabilityPort,
-    runtime: &R,
+struct ReadRuntimeContext {
     actor: AgentName,
     actor_team: Option<TeamName>,
     target: crate::mailbox::source::ResolvedTarget,
     seen_watermark: Option<IsoTimestamp>,
-) -> Result<ReadOutcome, AtmError> {
-    let own_inbox = actor == target.agent && actor_team.as_deref() == Some(target.team.as_str());
-    let workflow_path =
-        runtime.workflow_state_path(&query.home_dir, &target.team, &target.agent)?;
-    let mut workflow_state =
-        runtime.load_workflow_state(&query.home_dir, &target.team, &target.agent)?;
-    let mut source_files =
-        runtime.observe_source_files(&query.home_dir, &target.team, &target.agent)?;
-    let (mut bucket_counts, mut selected) =
-        selection_state_for_source_files(&source_files, &workflow_state, &query, seen_watermark);
-    let mut timed_out = false;
+}
 
-    if selected.is_empty()
-        && let Some(timeout_secs) = query.timeout_secs
-    {
-        let wait_satisfied = wait::wait_for_eligible_message(
-            timeout_secs,
-            || {
-                Ok(apply_idle_notification_dedup(
-                    dedupe_message_id_surface(
-                        merged_surface(&runtime.observe_source_files(
-                            &query.home_dir,
-                            &target.team,
-                            &target.agent,
-                        )?),
-                        |message: &SourcedMessage| message.envelope.message_id,
-                        |message: &SourcedMessage| message.envelope.timestamp,
-                    ),
-                    &workflow_state,
-                ))
-            },
-            |messages| {
-                !selected_after_filters(messages, &workflow_state, &query, seen_watermark)
-                    .is_empty()
-            },
-        )?;
+fn resolve_read_context<R: RetainedServiceRuntime>(
+    query: &ReadQuery,
+    runtime: &R,
+) -> Result<ReadRuntimeContext, AtmError> {
+    let config = runtime.load_config(&query.current_dir)?;
+    let actor = identity::resolve_actor_identity(query.actor_override.as_deref(), config.as_ref())?;
+    let actor_team = config::resolve_team(query.team_override.as_deref(), config.as_ref());
+    let target = resolve_target(
+        query.target_address.as_ref(),
+        &actor,
+        query.team_override.as_ref(),
+        config.as_ref(),
+    )?;
 
-        if wait_satisfied {
-            workflow_state =
-                runtime.load_workflow_state(&query.home_dir, &target.team, &target.agent)?;
-            source_files =
-                runtime.observe_source_files(&query.home_dir, &target.team, &target.agent)?;
-            (bucket_counts, selected) = selection_state_for_source_files(
-                &source_files,
-                &workflow_state,
-                &query,
-                seen_watermark,
-            );
-        } else {
-            timed_out = true;
-        }
+    let team_dir = runtime.team_dir(&query.home_dir, &target.team)?;
+    if !team_dir.exists() {
+        return Err(AtmError::team_not_found(&target.team).with_recovery(
+            "Create the team config for the requested team or target a different team before retrying `atm read`.",
+        ));
     }
 
-    let match_count = selected.len();
-    sort_and_limit_selected(&mut selected, Some(1));
-    let mutation_needed = displayed_messages_require_mutation(&selected);
-
-    let (mutation_applied, output_message, bucket_counts, selected_message_id, match_count) =
-        if timed_out || selected.is_empty() || !mutation_needed {
-            (
-                false,
-                output_messages_from_selection(&selected, &source_files, &workflow_state)
-                    .into_iter()
-                    .next(),
-                bucket_counts,
-                selected
-                    .first()
-                    .and_then(|message| message.envelope.message_id),
-                match_count,
-            )
-        } else {
-            runtime.with_locked_source_files(
-                &query.home_dir,
-                &target.team,
-                &target.agent,
-                [workflow_path],
-                runtime.mailbox_timeout_policy().workflow_lock_timeout,
-                |_source_paths, source_files| {
-                    let mut workflow_state = runtime.load_workflow_state(
-                        &query.home_dir,
-                        &target.team,
-                        &target.agent,
-                    )?;
-                    let (bucket_counts, mut selected) = selection_state_for_source_files(
-                        source_files,
-                        &workflow_state,
-                        &query,
-                        seen_watermark,
-                    );
-                    let match_count = selected.len();
-                    sort_and_limit_selected(&mut selected, Some(1));
-                    let mutation = apply_display_mutations(
-                        source_files,
-                        &mut workflow_state,
-                        &selected,
-                        query.ack_activation_mode,
-                        own_inbox,
-                    );
-                    if mutation.mailbox_changed {
-                        runtime.commit_source_files(source_files)?;
-                    }
-                    if mutation.workflow_changed {
-                        runtime.save_workflow_state(
-                            &query.home_dir,
-                            &target.team,
-                            &target.agent,
-                            &workflow_state,
-                        )?;
-                    }
-                    let output_message =
-                        output_messages_from_selection(&selected, source_files, &workflow_state)
-                            .into_iter()
-                            .next();
-                    Ok((
-                        mutation.any_changed,
-                        output_message,
-                        bucket_counts,
-                        selected
-                            .first()
-                            .and_then(|message| message.envelope.message_id),
-                        match_count,
-                    ))
-                },
-            )?
-        };
-
-    if query.seen_state_update
-        && !selected.is_empty()
-        && let Some(latest_timestamp) = selected
+    let team_config = runtime.load_team_config(&team_dir)?;
+    if target.explicit
+        && !team_config
+            .members
             .iter()
-            .map(|message| message.envelope.timestamp)
-            .max()
+            .any(|member| member.name == target.agent.as_str())
     {
-        runtime.save_seen_watermark(
-            &query.home_dir,
-            &target.team,
-            &target.agent,
-            latest_timestamp,
-        )?;
-    }
-
-    let outcome = ReadOutcome {
-        action: CommandAction::Read,
-        team: target.team.clone(),
-        agent: target.agent.clone(),
-        selection_mode: query.selection_mode,
-        mutation_applied,
-        count: usize::from(output_message.is_some()),
-        message: output_message,
-        selected_message_id,
-        match_count,
-        additional_match_count: match_count.saturating_sub(1),
-        bucket_counts,
-    };
-
-    let _ = observability.emit(CommandEvent {
-        command: "read",
-        action: "read",
-        outcome: if timed_out { "timeout" } else { "ok" },
-        team: outcome.team.clone(),
-        agent: outcome.agent.clone(),
-        sender: actor,
-        message_id: None,
-        requires_ack: false,
-        dry_run: false,
-        task_id: None,
-        error_code: None,
-        error_message: None,
-    });
-
-    Ok(outcome)
-}
-
-pub(crate) fn selection_state_for_source_files(
-    source_files: &[SourceFile],
-    workflow_state: &workflow::WorkflowStateFile,
-    query: &ReadQuery,
-    seen_watermark: Option<IsoTimestamp>,
-) -> (BucketCounts, Vec<ClassifiedMessage>) {
-    let classified_all = classify_all(
-        apply_idle_notification_dedup(
-            dedupe_message_id_surface(
-                merged_surface(source_files),
-                |message: &SourcedMessage| message.envelope.message_id,
-                |message: &SourcedMessage| message.envelope.timestamp,
+        return Err(
+            AtmError::agent_not_found(&target.agent, &target.team).with_recovery(
+                "Update the team membership in config.json or read a different mailbox target.",
             ),
-            workflow_state,
-        ),
-        workflow_state,
-    );
-    let logical_current = logical_current_messages(classified_all.clone());
-    let bucket_counts = bucket_counts_for(&logical_current);
-    if let Some(message_id) = query.message_id_filter {
-        let selected = classified_all
-            .into_iter()
-            .filter(|message| message.envelope.message_id == Some(message_id))
-            .collect();
-        return (bucket_counts, selected);
-    }
-    let filtered = apply_filters(
-        logical_current,
-        query.sender_filter.as_ref(),
-        query.timestamp_filter,
-        query.task_filter.as_ref(),
-        query.contains_filter.as_deref(),
-    );
-    let selected = select_messages(&filtered, query.selection_mode, seen_watermark);
-    (bucket_counts, selected)
-}
-
-fn selection_state_for_mailbox_metadata_rows(
-    rows: &[boundary::MailStoreMailboxMetadataRow],
-    query: &ReadQuery,
-    seen_watermark: Option<IsoTimestamp>,
-) -> (BucketCounts, Vec<ClassifiedMessage>) {
-    let classified_all = classify_mailbox_metadata_rows(rows);
-    let logical_current = logical_current_messages(classified_all.clone());
-    let bucket_counts = bucket_counts_for(&logical_current);
-    if let Some(message_id) = query.message_id_filter {
-        let selected = classified_all
-            .into_iter()
-            .filter(|message| message.envelope.message_id == Some(message_id))
-            .collect();
-        return (bucket_counts, selected);
-    }
-    let filtered = apply_filters(
-        logical_current,
-        query.sender_filter.as_ref(),
-        query.timestamp_filter,
-        query.task_filter.as_ref(),
-        query.contains_filter.as_deref(),
-    );
-    let selected = select_messages(&filtered, query.selection_mode, seen_watermark);
-    (bucket_counts, selected)
-}
-
-fn classify_mailbox_metadata_rows(
-    rows: &[boundary::MailStoreMailboxMetadataRow],
-) -> Vec<ClassifiedMessage> {
-    let projected = rows
-        .iter()
-        .enumerate()
-        .map(|(index, row)| ClassifiedMessage {
-            source_index: index.into(),
-            source_path: PathBuf::from(row.message_key.as_ref()),
-            bucket: DisplayBucket::Unread,
-            class: MessageClass::Unread,
-            envelope: MessageEnvelope {
-                from: row.from_agent.clone(),
-                text: row.summary.clone().unwrap_or_default(),
-                timestamp: row.message_at,
-                read: row.read,
-                source_team: None,
-                summary: row.summary.clone(),
-                message_id: row.message_id,
-                pending_ack_at: row.pending_ack.then_some(row.message_at),
-                acknowledged_at: row.acknowledged_at,
-                acknowledges_message_id: None,
-                parent_message_id: row.parent_message_id,
-                thread_mode: row.thread_mode,
-                expires_at: None,
-                task_id: row.task_id.clone(),
-                extra: serde_json::Map::new(),
-            },
-        })
-        .collect::<Vec<_>>();
-    let envelopes = projected
-        .iter()
-        .map(|message| message.envelope.clone())
-        .collect::<Vec<_>>();
-    let thread_index = ThreadIndex::new(&envelopes);
-
-    projected
-        .into_iter()
-        .map(|message| {
-            let effective = effective_display_envelope(&message.envelope, &thread_index);
-            let class = state::classify_message(&effective);
-            let bucket = state::display_bucket_for_class(class);
-            ClassifiedMessage {
-                source_index: message.source_index,
-                source_path: message.source_path,
-                bucket,
-                class,
-                envelope: effective,
-            }
-        })
-        .collect()
-}
-
-fn merged_surface(source_files: &[SourceFile]) -> Vec<SourcedMessage> {
-    source_files
-        .iter()
-        .flat_map(|source| {
-            source
-                .messages
-                .iter()
-                .cloned()
-                .enumerate()
-                .map(|(source_index, envelope)| SourcedMessage {
-                    envelope,
-                    source_path: source.path.clone(),
-                    source_index: source_index.into(),
-                })
-        })
-        .collect()
-}
-
-fn apply_idle_notification_dedup(
-    deduped: Vec<SourcedMessage>,
-    workflow_state: &workflow::WorkflowStateFile,
-) -> Vec<SourcedMessage> {
-    let projected = deduped
-        .into_iter()
-        .map(|message| SourcedMessage {
-            envelope: workflow::project_envelope(&message.envelope, workflow_state),
-            source_path: message.source_path,
-            source_index: message.source_index,
-        })
-        .collect::<Vec<_>>();
-    let latest_idle_for_sender = messages_from_idle_sender(&projected);
-
-    projected
-        .into_iter()
-        .enumerate()
-        .filter_map(|(index, message)| {
-            dedupe_idle_notifications(index, &message, &latest_idle_for_sender).then_some(message)
-        })
-        .collect()
-}
-
-fn dedupe_idle_notifications(
-    index: usize,
-    message: &SourcedMessage,
-    latest_idle_for_sender: &HashMap<AgentName, usize>,
-) -> bool {
-    if !is_unread_idle_notification(&message.envelope) {
-        return true;
+        );
     }
 
-    idle_sender(&message.envelope)
-        .and_then(|sender| latest_idle_for_sender.get(&sender))
-        .map(|keep_index| *keep_index == index)
-        .unwrap_or(true)
-}
-
-fn messages_from_idle_sender(messages: &[SourcedMessage]) -> HashMap<AgentName, usize> {
-    let mut latest_idle_for_sender = HashMap::new();
-
-    for (index, message) in messages.iter().enumerate() {
-        if !is_unread_idle_notification(&message.envelope) {
-            continue;
-        }
-
-        if let Some(sender) = idle_sender(&message.envelope) {
-            latest_idle_for_sender
-                .entry(sender)
-                .and_modify(|keep_index| *keep_index = index)
-                .or_insert(index);
-        }
-    }
-
-    latest_idle_for_sender
-}
-
-fn is_unread_idle_notification(message: &MessageEnvelope) -> bool {
-    !message.read && idle_notification_sender(message).is_some()
-}
-
-fn idle_sender(message: &MessageEnvelope) -> Option<AgentName> {
-    idle_notification_sender(message)
-}
-
-fn idle_notification_sender(message: &MessageEnvelope) -> Option<AgentName> {
-    let value = match serde_json::from_str::<Value>(&message.text) {
-        Ok(value) => value,
-        Err(error) => {
-            if message.text.contains("idle_notification") {
-                debug!(
-                    %error,
-                    recovery = "Repair or remove the malformed Claude idle-notification JSON. ATM will continue treating the record as a normal mailbox message.",
-                    message_text = %message.text,
-                    "ignoring malformed idle-notification JSON while classifying read surface"
-                );
-            }
-            return None;
-        }
-    };
-
-    if value.get("type").and_then(Value::as_str) != Some("idle_notification") {
-        return None;
-    }
-
-    match value.get("from").and_then(Value::as_str) {
-        Some(sender) => match sender.parse() {
-            Ok(sender) => Some(sender),
-            Err(error) => {
-                debug!(
-                    %error,
-                    recovery = "Ensure Claude idle-notification payloads include a valid ATM agent name in `from`. ATM will continue treating the record as a normal mailbox message.",
-                    sender,
-                    message_text = %message.text,
-                    "ignoring malformed idle-notification payload with invalid `from`"
-                );
-                None
-            }
-        },
-        None => {
-            debug!(
-                recovery = "Ensure Claude idle-notification payloads include a string `from` field. ATM will continue treating the record as a normal mailbox message.",
-                message_text = %message.text,
-                "ignoring malformed idle-notification payload missing string `from`"
-            );
-            None
-        }
-    }
-}
-
-fn classify_all(
-    messages: Vec<SourcedMessage>,
-    workflow_state: &workflow::WorkflowStateFile,
-) -> Vec<ClassifiedMessage> {
-    let projected = messages
-        .iter()
-        .map(|message| workflow::project_envelope(&message.envelope, workflow_state))
-        .collect::<Vec<_>>();
-    let thread_index = ThreadIndex::new(&projected);
-
-    messages
-        .into_iter()
-        .zip(projected.iter().cloned())
-        .map(|(message, projected)| {
-            let effective = effective_display_envelope(&projected, &thread_index);
-            let class = state::classify_message(&effective);
-            let bucket = state::display_bucket_for_class(class);
-
-            ClassifiedMessage {
-                source_index: message.source_index,
-                source_path: message.source_path,
-                bucket,
-                class,
-                envelope: effective,
-            }
-        })
-        .collect()
-}
-
-fn apply_filters(
-    messages: Vec<ClassifiedMessage>,
-    sender_filter: Option<&AgentName>,
-    timestamp_filter: Option<IsoTimestamp>,
-    task_filter: Option<&TaskId>,
-    contains_filter: Option<&str>,
-) -> Vec<ClassifiedMessage> {
-    filters::apply_contains_filter(
-        filters::apply_task_filter(
-            filters::apply_timestamp_filter(
-                filters::apply_sender_filter(messages, sender_filter),
-                timestamp_filter,
-            ),
-            task_filter,
-        ),
-        contains_filter,
-    )
-}
-
-fn logical_current_messages(messages: Vec<ClassifiedMessage>) -> Vec<ClassifiedMessage> {
-    let projected = messages
-        .iter()
-        .map(|message| message.envelope.clone())
-        .collect::<Vec<_>>();
-    let thread_index = ThreadIndex::new(&projected);
-
-    messages
-        .into_iter()
-        .filter(|message| {
-            message
-                .envelope
-                .message_id
-                .is_none_or(|message_id| thread_index.is_terminal(message_id))
-        })
-        .map(|mut message| {
-            if let Some(message_id) = message.envelope.message_id
-                && let Some(logical) = thread_index.logical_current_envelope(message_id)
-            {
-                message.envelope = logical;
-            }
-            message
-        })
-        .collect()
-}
-
-fn bucket_counts_for(messages: &[ClassifiedMessage]) -> BucketCounts {
-    messages.iter().fold(
-        BucketCounts {
-            unread: 0,
-            pending_ack: 0,
-            history: 0,
-        },
-        |mut counts, message| {
-            if hidden_from_normal_views(&message.envelope) {
-                return counts;
-            }
-            match message.bucket {
-                DisplayBucket::Unread => counts.unread += 1,
-                DisplayBucket::PendingAck => counts.pending_ack += 1,
-                DisplayBucket::History => counts.history += 1,
-            }
-            counts
-        },
-    )
-}
-
-fn select_messages(
-    messages: &[ClassifiedMessage],
-    selection_mode: ReadSelection,
-    seen_watermark: Option<IsoTimestamp>,
-) -> Vec<ClassifiedMessage> {
-    let watermark = if selection_mode == ReadSelection::All {
-        None
+    let seen_watermark = if query.seen_state_filter && query.selection_mode != ReadSelection::All {
+        runtime.load_seen_watermark(&query.home_dir, &target.team, &target.agent)?
     } else {
-        seen_watermark
+        None
     };
 
-    let visible = messages
-        .iter()
-        .filter(|message| !hidden_for_selection(&message.envelope, selection_mode))
-        .cloned()
-        .collect::<Vec<_>>();
-
-    filters::apply_selection_mode(visible, selection_mode, watermark)
-}
-
-fn selected_after_filters(
-    messages: &[SourcedMessage],
-    workflow_state: &workflow::WorkflowStateFile,
-    query: &ReadQuery,
-    seen_watermark: Option<IsoTimestamp>,
-) -> Vec<ClassifiedMessage> {
-    let classified = classify_all(messages.to_vec(), workflow_state);
-    if let Some(message_id) = query.message_id_filter {
-        return classified
-            .into_iter()
-            .filter(|message| message.envelope.message_id == Some(message_id))
-            .collect();
-    }
-    let filtered = apply_filters(
-        logical_current_messages(classified),
-        query.sender_filter.as_ref(),
-        query.timestamp_filter,
-        query.task_filter.as_ref(),
-        query.contains_filter.as_deref(),
-    );
-    select_messages(&filtered, query.selection_mode, seen_watermark)
-}
-
-pub(crate) fn sort_and_limit_selected(selected: &mut Vec<ClassifiedMessage>, limit: Option<usize>) {
-    selected.sort_by(|left, right| {
-        right
-            .envelope
-            .timestamp
-            .cmp(&left.envelope.timestamp)
-            .then_with(|| right.envelope.message_id.cmp(&left.envelope.message_id))
-            .then_with(|| right.source_index.cmp(&left.source_index))
-    });
-
-    if let Some(limit) = limit {
-        selected.truncate(limit);
-    }
+    Ok(ReadRuntimeContext {
+        actor,
+        actor_team,
+        target,
+        seen_watermark,
+    })
 }
 
 fn message_key_for_classified(
@@ -1200,36 +649,6 @@ fn output_messages_from_selection(
         .collect()
 }
 
-fn effective_display_envelope(
-    envelope: &MessageEnvelope,
-    thread_index: &ThreadIndex<'_>,
-) -> MessageEnvelope {
-    let Some(message_id) = envelope.message_id else {
-        return envelope.clone();
-    };
-    if thread_index.is_terminal(message_id) {
-        return envelope.clone();
-    }
-
-    let mut historical = envelope.clone();
-    historical.read = true;
-    historical.pending_ack_at = None;
-    historical
-}
-
-fn hidden_from_normal_views(envelope: &MessageEnvelope) -> bool {
-    let now = IsoTimestamp::now();
-    is_expired_ephemeral(envelope, now) || (is_ephemeral(envelope) && envelope.read)
-}
-
-fn hidden_for_selection(envelope: &MessageEnvelope, selection_mode: ReadSelection) -> bool {
-    let now = IsoTimestamp::now();
-    if is_expired_ephemeral(envelope, now) {
-        return true;
-    }
-    selection_mode != ReadSelection::All && is_ephemeral(envelope) && envelope.read
-}
-
 #[derive(Debug, Default, Clone, Copy)]
 struct DisplayMutationResult {
     any_changed: bool,
@@ -1334,9 +753,9 @@ mod tests {
     use serde_json::Map;
     use tempfile::tempdir;
 
-    use super::{
-        ReadQuery, idle_notification_sender, selected_after_filters,
-        selection_state_for_source_files,
+    use super::ReadQuery;
+    use super::legacy_path::{
+        idle_notification_sender, selected_after_filters, selection_state_for_source_files,
     };
     use crate::mailbox::source::SourceFile;
     use crate::mailbox::source::SourcedMessage;
