+++ docs/atm-rusqlite/sql_load-message.mmd
@@ -1,9 +1,9 @@
 %% title: `Load Message`
 %% summary: SQLite boundary method `MailStore::load_message`.
-%% commentary: This SQL-boundary path shows the direct immutable-row load. It currently reads only `mail_messages`, rebuilds one `MailStoreMessageRecord`, and returns `None` when no row exists.
-%% read_tables: mail_messages
+%% commentary: This SQL-boundary path reads the immutable row from `mail_messages`, LEFT JOINs `mail_message_states`, applies the loaded mutable state projection to the envelope, and returns `None` when no row exists.
+%% read_tables: mail_messages, mail_message_states
 %% write_tables:
-%% notes: reads: 1 | writes: 0 | joins: 0 | transactions: 0
+%% notes: reads: 1 | writes: 0 | joins: 1 | transactions: 0
 %% sets: query
 flowchart TD
     A["<b>Load Message</b><br/>MailStore::load_message"]
@@ -11,13 +11,18 @@ flowchart TD
 team
 agent
 message_key"]
-    I --> R["Read message row
-SELECT envelope_json, imported_from, recorded_at
+    I --> R["Read message row + state
+SELECT envelope_json, imported_from, recorded_at,
+       state.read, state.pending_ack_at,
+       state.acknowledged_at, state.expires_at,
+       state.deleted_at, state.updated_at
 FROM mail_messages
+LEFT JOIN mail_message_states ...
 WHERE team, agent, message_key"]
     R -->|row found| M["Materialize record
 deserialize envelope_json
 parse recorded_at?
+apply_loaded_state
 rebuild MailStoreMessageRecord"]
     R -->|no row| O0["Output
 record: None"]
