+++ docs/legacy-atm-message-schema.md
@@ -33,7 +33,7 @@ Observed historical ATM-only fields used for alerts and repair notices:
 
 These fields are accepted for backward compatibility with historical inbox
 data. They are not the forward schema contract for newly-authored ATM machine
-metadata, which belongs under `metadata.atm` in
+state, which now lives in SQLite as documented in
 [`atm-message-schema.md`](./atm-message-schema.md).
 
 ## 3. Read Compatibility Rule
@@ -43,7 +43,7 @@ ATM read and related workflows must continue to accept:
 - Claude Code-native messages
 - historical ATM messages using the legacy top-level additive fields documented
   here
-- future ATM metadata-based messages documented in
+- current ATM-authored compatibility messages documented in
   [`atm-message-schema.md`](./atm-message-schema.md)
 
 ## 4. Write Deprecation Rule
@@ -52,8 +52,8 @@ This schema is deprecated for write:
 
 - ATM must not introduce new ATM-only top-level fields under this schema
 - existing historical fields remain readable
-- migration to metadata-based ATM machine fields is documented in
+- forward ATM machine state is documented as SQLite-owned in
   [`atm-message-schema.md`](./atm-message-schema.md)
 - legacy top-level `atmAlertKind` and `missingConfigPath` remain read-compatible
-  during the migration period and must not be removed from compatibility docs
-  before the forward metadata migration is complete
+  during the cleanup period and must not be removed from compatibility docs
+  before the surviving read-compatibility paths are retired
